requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlmodel>=0.0.14
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.0.0
email-validator>=2.0.0
